Conflicts:
	.travis.yml
	build_client.sh

+++ build_gemstone.sh
@@ -7,10 +7,12 @@
 
 # help function
 function display_help() {
-	echo "$(basename $0) -i input -o output {-m} {-s script} {-f full-path-to-script} {-G} {-X}"
+	echo "$(basename $0) -i input -o output {-m} {-n} {-s script} {-d} {-f full-path-to-script} {-G} {-X}"
+  echo " -d skip delete of OUTPUT_PATH"
 	echo " -f one or more scripts (full path) to build the image, can be intermixed with -m and -s options"
 	echo " -i input product name, image from images-directory, or successful jenkins build"
 	echo " -m use Metacello test harness: FileTree, Metacello, travisCIHarness.st, can be intermixed with -f and -s options"
+  echo " -n start a netldi"
 	echo " -o output product name"
 	echo " -s one or more scripts from the scripts-directory to build the image, can be intermixed with -m and -f options"
 	echo " -G do not bootstrap GemStone to a known version of GLASS (currently GLASS 1.0-beta.8.7.4). Implies -X"
@@ -22,10 +24,15 @@ echo "PROCESSING OPTIONS"
 # parse options
 BOOTSTRAP_METACELLO="include"
 BOOTSTRAP_GLASS="include"
-while getopts ":i:mGXo:f:s:?" OPT ; do
+DELETE_OUTPUT_PATH=true
+while getopts ":i:mnGXo:f:s:?" OPT ; do
 	case "$OPT" in
 
-		# full path to script
+		    # skip delete of OUTPUT_PATH
+    d) DELETE_OUTPUT_PATH=false
+    ;;
+
+      # full path to script
 	  	f)	if [ -f "$OPTARG" ] ; then
                 	SCRIPTS=("${SCRIPTS[@]}" "$OPTARG")
 			else
@@ -39,9 +46,16 @@ while getopts ":i:mGXo:f:s:?" OPT ; do
 			GEMSTONE_VERSION="$OPTARG"
 			;;
 
-    		# include standard Metacello test harness
-    		m) SCRIPTS=("${SCRIPTS[@]}" "$SCRIPTS_PATH/travisCIHarness.st" )
-    			;; 
+    # include standard Metacello test harness
+   	m) SCRIPTS=("${SCRIPTS[@]}" "$SCRIPTS_PATH/travisCIHarness.st" )
+   		;;
+
+    # netldi
+    n)
+      source /opt/gemstone/product/seaside/defSeaside #set GemStone environment variables
+      startnet
+      gslist -lc
+      ;;
 
 		# output
 		o)	OUTPUT_NAME="$OPTARG"
@@ -79,8 +93,8 @@ while getopts ":i:mGXo:f:s:?" OPT ; do
 			;;
 
 		# exclude Metacello bootstrap
-    		X) BOOTSTRAP_METACELLO="exclude"
-    			;;
+    X) BOOTSTRAP_METACELLO="exclude"
+    	;;
 
 		# show help
 		\?)	display_help
@@ -88,14 +102,18 @@ while getopts ":i:mGXo:f:s:?" OPT ; do
 		;;
 
 	esac
+
 done
 
 echo "preparing script files"
 
-# prepare output path
-if [ -d "$OUTPUT_PATH" ] ; then
-	rm -rf "$OUTPUT_PATH"
+#prepare output path
+if [ "$DELETE_OUTPUT_PATH" == true ] ; then
+  if [ -d "$OUTPUT_PATH" ] ; then
+	  rm -rf "$OUTPUT_PATH"
+  fi
 fi
+
 mkdir -p "$OUTPUT_PATH"
 
 # hook up the git_cache, Metacello bootstrap and mcz repo
@@ -109,17 +127,17 @@ ln -sf "$BUILDER_CI_HOME/scripts/MetacelloBuilderTravisCI.st" "$OUTPUT_PATH/"
 if [ "$BOOTSTRAP_GLASS" == "include" ] ; then
 	BEFORE_SCRIPTS=("$SCRIPTS_PATH/before_gemstone.st" "$SCRIPTS_PATH/before.st")
 else
-        BEFORE_SCRIPTS=("$SCRIPTS_PATH/patch_gemstone.st" "$SCRIPTS_PATH/before.st")
+  BEFORE_SCRIPTS=("$SCRIPTS_PATH/patch_gemstone.st" "$SCRIPTS_PATH/before.st")
 fi
 
 
 # special doit needed for 2.4.4.x
 case "$ST" in
-        GemStone-2.4.4.1|GemStone-2.4.4.7)
-	BEFORE_SCRIPTS=("${BEFORE_SCRIPTS[@]}" "$SCRIPTS_PATH/gemstone244x.st")
-        ;;
-        *) #do nothing by default
-        ;;
+  GemStone-2.4.4.1|GemStone-2.4.4.7)
+	   BEFORE_SCRIPTS=("${BEFORE_SCRIPTS[@]}" "$SCRIPTS_PATH/gemstone244x.st")
+     ;;
+  *) #do nothing by default
+     ;;
 esac
 
 # prepare script file
@@ -138,9 +156,8 @@ for FILE in "${SCRIPTS[@]}" ; do
 	echo "commit" >> "$OUTPUT_SCRIPT"
 done
 
-cd ${BUILD_PATH}/travisCI
+cd ${BUILD_PATH}/${OUTPUT_NAME}
 source /opt/gemstone/product/seaside/defSeaside #set GemStone environment variables
-# gslist -lc
 
 echo "synchronize timezones"
